backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
sentence-transformers==2.3.1
chromadb==0.4.22
mtcnn==0.1.1
tensorflow==2.15.0
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
supabase==2.3.4
python-dotenv==1.0.0
